Refactor model rendering logic into separate class
