+++ core/src/main/java/org/eclipse/epsilon/labs/playground/fn/flexmi2plantuml/Flexmi2PlantUMLRequest.java
@@ -3,7 +3,7 @@ package org.eclipse.epsilon.labs.playground.fn.flexmi2plantuml;
 import io.micronaut.serde.annotation.Serdeable;
 
 @Serdeable
-public class FlexmiToPlantUMLRequest {
+public class Flexmi2PlantUMLRequest {
 
     private String flexmi;
     private String emfatic;
