+++ core/src/test/java/org/eclipse/epsilon/labs/playground/fn/flexmi2plantuml/Flexmi2PlantUMLClient.java
@@ -1,5 +1,7 @@
 package org.eclipse.epsilon.labs.playground.fn.flexmi2plantuml;
 
+import org.eclipse.epsilon.labs.playground.fn.ModelDiagramResponse;
+
 import io.micronaut.http.annotation.Body;
 import io.micronaut.http.annotation.Post;
 import io.micronaut.http.client.annotation.Client;
@@ -7,5 +9,5 @@ import io.micronaut.http.client.annotation.Client;
 @Client(Flexmi2PlantUMLController.PATH)
 public interface Flexmi2PlantUMLClient {
     @Post("/")
-    ModelDiagramResponse convert(@Body FlexmiToPlantUMLRequest request);
+    ModelDiagramResponse convert(@Body Flexmi2PlantUMLRequest request);
 }
