+++ core/src/test/java/org/eclipse/epsilon/labs/playground/fn/xmi2plantuml/Xmi2PlantUMLClient.java
@@ -1,5 +1,7 @@
 package org.eclipse.epsilon.labs.playground.fn.xmi2plantuml;
 
+import org.eclipse.epsilon.labs.playground.fn.ModelDiagramResponse;
+
 import io.micronaut.http.annotation.Body;
 import io.micronaut.http.annotation.Post;
 import io.micronaut.http.client.annotation.Client;
@@ -7,5 +9,5 @@ import io.micronaut.http.client.annotation.Client;
 @Client(Xmi2PlantUMLController.PATH)
 public interface Xmi2PlantUMLClient {
     @Post("/")
-    ModelDiagramResponse convert(@Body XmiToPlantUMLRequest request);
+    ModelDiagramResponse convert(@Body Xmi2PlantUMLRequest request);
 }
